Split expected symbol leaf tests
